Domain Maker/requirements.txt
python-dotenv
pytrends
producthunt
beautifulsoup4
requests
pandas
scikit-learn
openai


